+++ Corso/libft/ft_memset.c
@@ -6,12 +6,13 @@
 /*   By: mcoppola <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2023/01/17 13:09:22 by mcoppola          #+#    #+#             */
-/*   Updated: 2023/01/18 17:54:32 by mcoppola         ###   ########.fr       */
+/*   Updated: 2023/01/20 14:34:40 by mcoppola         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
+#include "libft.h"
 // potrebbe non dover fare il controllo ((char *)s)[i] all'interno del while
-void	*ft_memset(void *s, int c, int n)
+void	*ft_memset(void *s, int c, size_t n)
 {
 	int	i;
